+++ src/block-patterns/homepage-posts-19.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * Homepage Posts Pattern 19.
+ * Content Loop Pattern 19.
  *
  * @package Newspack_Blocks
  */
 
 return array(
-	'title'         => __( 'Homepage Posts Pattern 19', 'newspack-blocks' ),
+	'title'         => __( 'Content Loop Pattern 19', 'newspack-blocks' ),
 	'content'       => '<!-- wp:newspack-blocks/homepage-articles {"className":"is-style-default","showExcerpt":false,"imageShape":"portrait","minHeight":50,"showAuthor":false,"postLayout":"grid","columns":4,"postsToShow":4,"mediaPosition":"behind"} /-->',
 	'viewportWidth' => 1280,
 	'categories'    => array( 'newspack-homepage-posts' ),