+++ src/block-patterns/homepage-posts-24.php
@@ -1,12 +1,12 @@
 <?php
 /**
- * Homepage Posts Pattern 24.
+ * Content Loop Pattern 24.
  *
  * @package Newspack_Blocks
  */
 
 return array(
-	'title'         => __( 'Homepage Posts Pattern 24', 'newspack-blocks' ),
+	'title'         => __( 'Content Loop Pattern 24', 'newspack-blocks' ),
 	'content'       => "<!-- wp:columns -->\n<div class=\"wp-block-columns\"><!-- wp:column {\"width\":50} -->\n<div class=\"wp-block-column\" style=\"flex-basis:50%\"><!-- wp:newspack-blocks/homepage-articles {\"excerptLength\":50,\"showAvatar\":false,\"postsToShow\":1,\"typeScale\":5} /--></div>\n<!-- /wp:column -->\n\n<!-- wp:column {\"width\":25} -->\n<div class=\"wp-block-column\" style=\"flex-basis:25%\"><!-- wp:newspack-blocks/homepage-articles {\"showExcerpt\":false,\"excerptLength\":20,\"showAvatar\":false,\"postsToShow\":2,\"typeScale\":3} /--></div>\n<!-- /wp:column -->\n\n<!-- wp:column {\"width\":25} -->\n<div class=\"wp-block-column\" style=\"flex-basis:25%\"><!-- wp:newspack-blocks/homepage-articles {\"className\":\"is-style-default\",\"showExcerpt\":false,\"excerptLength\":15,\"showAvatar\":false,\"postsToShow\":2,\"typeScale\":3} /--></div>\n<!-- /wp:column --></div>\n<!-- /wp:columns -->",
 	'viewportWidth' => 1000,
 	'categories'    => array( 'newspack-homepage-posts' ),