+++ src/blocks/carousel/index.js
@@ -14,9 +14,13 @@ import edit from './edit';
  */
 import './view.scss';
 import './editor.scss';
+import metadata from './block.json';
+const { name, attributes, category } = metadata;
 
-export const name = 'carousel';
-export const title = __( 'Post Carousel' );
+// Name must be exported separately.
+export { name };
+
+export const title = __( 'Content Carousel', 'newspack-blocks' );
 
 export const icon = (
 	<SVG xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24">
@@ -33,109 +37,20 @@ export const settings = {
 	title,
 	icon: {
 		src: icon,
-		foreground: '#36f',
+		foreground: '#406ebc',
 	},
-	category: 'newspack',
-	keywords: [ __( 'posts' ), __( 'slideshow' ), __( 'carousel' ) ],
-	description: __( 'A carousel of posts.' ),
-	attributes: {
-		className: {
-			type: 'string',
-		},
-		imageFit: {
-			type: 'string',
-			default: 'cover',
-		},
-		autoplay: {
-			type: 'boolean',
-			default: false,
-		},
-		delay: {
-			type: 'number',
-			default: 5,
-		},
-		postsToShow: {
-			type: 'integer',
-			default: 3,
-		},
-		authors: {
-			type: 'array',
-		},
-		categories: {
-			type: 'array',
-		},
-		includeSubcategories: {
-			type: 'boolean',
-			default: true,
-		},
-		tags: {
-			type: 'array',
-		},
-		customTaxonomies: {
-			type: 'array',
-		},
-		showDate: {
-			type: 'boolean',
-			default: true,
-		},
-		showAuthor: {
-			type: 'boolean',
-			default: true,
-		},
-		showAvatar: {
-			type: 'boolean',
-			default: true,
-		},
-		showCaption: {
-			type: 'boolean',
-			default: false,
-		},
-		showCredit: {
-			type: 'boolean',
-			default: false,
-		},
-		showCategory: {
-			type: 'boolean',
-			default: false,
-		},
-		showTitle: {
-			type: 'boolean',
-			default: true,
-		},
-		postType: {
-			type: 'array',
-			default: [ 'post' ],
-			items: {
-				type: 'string',
-			},
-		},
-		specificMode: {
-			type: 'boolean',
-			default: false,
-		},
-		specificPosts: {
-			type: 'array',
-			default: [],
-			items: { type: 'integer' },
-		},
-		slidesPerView: {
-			type: 'number',
-			default: 1,
-		},
-		hideControls: {
-			type: 'boolean',
-			default: false,
-		},
-		aspectRatio: {
-			type: 'number',
-			default: 0.75,
-		},
-		includedPostStatuses: {
-			type: 'array',
-			default: [ 'publish' ],
-			items: { type: 'string' },
-		},
-	},
+	attributes,
+	category,
+	keywords: [
+		__( 'posts', 'newspack-blocks' ),
+		__( 'articles', 'newspack-blocks' ),
+		__( 'latest', 'newspack-blocks' ),
+		__( 'query', 'newspack-blocks' ),
+	],
+	description: __(
+		'An advanced block that displays content in a carousel format with customizable parameters and visual configurations.',
+		'newspack-blocks'
+	),
 	supports: {
 		html: false,
 		align: [ 'center', 'wide', 'full' ],